fixed memory leak

+++ mcmc_cuda/inc/mcmc.cuh
@@ -1,1 +1,1 @@
-void mcmc(double *u,double *y, double *theta,int N,int order,int num_samples,double *theta_0,double elim);
+void mcmc(double *u,double *y, double *theta,long int N,int order,int num_samples,double *theta_0,double elim);
